- add fuzzy search - Publishes view refactor
